Refactor : 파일 위치 조정
- utils 패키지를 생성하여 InputUtils.java와 PrintUtils.java를 모았습니다.
