+++ src/main/java/racingcar/Game.java
@@ -1,6 +1,8 @@
 package racingcar;
 
-import java.util.ArrayList;
+import Utils.InputUtils;
+import Utils.PrintUtils;
+
 import java.util.List;
 
 public class Game {
